dubdilla-location-fix: new mod from nexus
upstream description:

One of MW's quests has voiced dialogue that sends you to a totally
different location than where the quest actually takes place (and
completely contradicts the journal for that matter). Hiring Wes
Johnson (voice actor) to rectify this is not gonna happen without a bag
of money. Redoing the voice by someone else would be lame, right?

So I basically just put the cave entrance in a more logical location; so
that this glaring and immersion breaking inconsistency is fixed.
• Directions in the journal now correspond to the dialogue.
• Mount Assarnibibi will be marked on your map as soon as the quest begins.
• Removed any sign of the previous entrance in the Grazelands.
• No landscape edits were made.